ChildProcess: account for a system error when launching a process
When a non-existing command is attempted to run, both status and
signal will be null, which leads to ChildProcess module crashing.
Obviously, this is incorrect behavior.

The problem is more general than that though: any system error that
would result in failing to run a process would result in the module
crash.

Fix that by checking for such case.

Fixes: https://github.com/purescript-node/purescript-node-child-process/issues/65